Move email reminders to their own folder (#625)
