+++ config/sidekiq.yml
@@ -21,12 +21,12 @@
     enabled: <%= ENV["ENABLE_THE_GRAPH_JOBS"] %>
   send_token_launch_reminder:
     cron: "0 10 * * *" # Runs once every day at 10am
-    class: SendTokenLaunchReminderJob
-    enabled: <%= ENV["TOKEN_LAUNCH_REMINDER_DAYS"].to_i > 0 %>
+    class: EmailReminders::SendTokenLaunchReminderJob
+    enabled: <%= ENV["EMAIL_REMINDER_DAYS"].to_i > 0 %>
   send_token_purchase_reminder:
     cron: "0 10 * * *" # Runs once every day at 10am
-    class: SendTokenPurchaseReminderJob
-    enabled: <%= ENV["TOKEN_PURCHASE_REMINDER_DAYS"].to_i > 0 %>
+    class: EmailReminders::SendTokenPurchaseReminderJob
+    enabled: <%= ENV["EMAIL_REMINDER_DAYS"].to_i > 0 %>
   send_complete_profile_reminder:
     cron: "0 10 * * *" # runs once every day at 10am
     class: EmailReminders::SendCompleteProfileReminderJob
